+++ heylogs-ext-forgejo/src/main/java/nbbrd/heylogs/ext/forgejo/Forgejo.java
@@ -23,6 +23,9 @@ public final class Forgejo implements Forge {
             .parser(REQUEST, ForgejoRequestLink::parse, ForgejoRequestRef::parse)
             .parser(MENTION, ForgejoMentionLink::parse, ForgejoMentionRef::parse)
             .linkParser(COMPARE, ForgejoCompareLink::parse)
-            .messageFetcher(new ForgejoMessageFetcher())
+            .linkResolver(ISSUE, ForgejoIssueLink::resolve)
+            .linkResolver(REQUEST, ForgejoRequestLink::resolve)
+            .messageFetcher(ISSUE, ForgejoMessageFetcher.ISSUE)
+            .messageFetcher(REQUEST, ForgejoMessageFetcher.REQUEST)
             .build();
 }