+++ heylogs-ext-forgejo/src/main/java/nbbrd/heylogs/ext/forgejo/ForgejoIssueLink.java
@@ -9,6 +9,7 @@ import nbbrd.heylogs.spi.ForgeRef;
 import nbbrd.io.http.URLQueryBuilder;
 import org.jspecify.annotations.Nullable;
 
+import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.regex.Pattern;
 
@@ -34,6 +35,20 @@ class ForgejoIssueLink implements ForgeLink {
         checkPathItem(pathArray, 3, NUMBER_PATTERN);
 
         return new ForgejoIssueLink(baseOf(url), pathArray[0], pathArray[1], parseInt(pathArray[3]));
+    }
+
+    @StaticFactoryMethod
+    public static @NonNull ForgejoIssueLink resolve(@NonNull URL projectUrl, @NonNull CharSequence ref) {
+        try {
+            return parse(
+                    URLQueryBuilder.of(projectUrl)
+                            .path(ISSUES_KEYWORD)
+                            .path(String.valueOf(ForgejoIssueRef.parse(ref).getIssueNumber()))
+                            .build());
+        } catch (
+                MalformedURLException ex) {
+            throw new RuntimeException(ex);
+        }
     }
 
     @NonNull