+++ heylogs-ext-gitlab/src/main/java/nbbrd/heylogs/ext/gitlab/GitLabIssueLink.java
@@ -6,8 +6,10 @@ import nbbrd.design.RepresentableAs;
 import nbbrd.design.StaticFactoryMethod;
 import nbbrd.heylogs.spi.ForgeLink;
 import nbbrd.heylogs.spi.ForgeRef;
+import nbbrd.io.http.URLQueryBuilder;
 import org.jspecify.annotations.Nullable;
 
+import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.List;
 
@@ -22,6 +24,20 @@ class GitLabIssueLink implements ForgeLink {
     @StaticFactoryMethod
     public static @NonNull GitLabIssueLink parse(@NonNull URL url) {
         return parseLink(GitLabIssueLink::new, ISSUES_KEYWORD, NUMBER_PATTERN, Integer::parseInt, url);
+    }
+
+    @StaticFactoryMethod
+    public static @NonNull GitLabIssueLink resolve(@NonNull URL projectUrl, @NonNull CharSequence ref) {
+        try {
+            return parse(
+                    URLQueryBuilder
+                            .of(projectUrl)
+                            .path(ISSUES_KEYWORD)
+                            .path(String.valueOf(GitLabIssueRef.parse(ref).getNumber()))
+                            .build());
+        } catch (MalformedURLException ex) {
+            throw new RuntimeException(ex);
+        }
     }
 
     @NonNull