+++ heylogs-ext-gitlab/src/main/java/nbbrd/heylogs/ext/gitlab/GitLabRequestLink.java
@@ -6,8 +6,10 @@ import nbbrd.design.RepresentableAs;
 import nbbrd.design.StaticFactoryMethod;
 import nbbrd.heylogs.spi.ForgeLink;
 import nbbrd.heylogs.spi.ForgeRef;
+import nbbrd.io.http.URLQueryBuilder;
 import org.jspecify.annotations.Nullable;
 
+import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.List;
 
@@ -22,6 +24,19 @@ class GitLabRequestLink implements ForgeLink {
     @StaticFactoryMethod
     public static @NonNull GitLabRequestLink parse(@NonNull URL url) {
         return parseLink(GitLabRequestLink::new, MERGE_REQUEST_KEYWORD, NUMBER_PATTERN, Integer::parseInt, url);
+    }
+
+    @StaticFactoryMethod
+    public static @NonNull GitLabRequestLink resolve(@NonNull URL projectUrl, @NonNull CharSequence ref) {
+        try {
+            return parse(
+                    URLQueryBuilder.of(projectUrl)
+                            .path(MERGE_REQUEST_KEYWORD)
+                            .path(String.valueOf(GitLabRequestRef.parse(ref).getNumber()))
+                            .build());
+        } catch (MalformedURLException ex) {
+            throw new RuntimeException(ex);
+        }
     }
 
     @NonNull