+++ heylogs-ext-gitlab/src/test/java/nbbrd/heylogs/ext/gitlab/GitLabRequestLinkTest.java
@@ -40,4 +40,31 @@ class GitLabRequestLinkTest {
         }
     }
 
+    @Test
+    void testResolve() {
+        // Basic number reference
+        assertThat(GitLabRequestLink.resolve(urlOf("https://gitlab.com/nbbrd/heylogs-ext-gitlab"), "!1"))
+                .returns(urlOf("https://gitlab.com"), GitLabRequestLink::getBase)
+                .returns(java.util.Arrays.asList("nbbrd"), GitLabRequestLink::getNamespace)
+                .returns("heylogs-ext-gitlab", GitLabRequestLink::getProject)
+                .returns(1, GitLabRequestLink::getNumber)
+                .hasToString("https://gitlab.com/nbbrd/heylogs-ext-gitlab/-/merge_requests/1");
+
+        // Namespace/project reference
+        assertThat(GitLabRequestLink.resolve(urlOf("https://gitlab.com/otherns/otherproj"), "nbbrd/heylogs-ext-gitlab!1"))
+                .returns(urlOf("https://gitlab.com"), GitLabRequestLink::getBase)
+                .returns(java.util.Arrays.asList("otherns"), GitLabRequestLink::getNamespace)
+                .returns("otherproj", GitLabRequestLink::getProject)
+                .returns(1, GitLabRequestLink::getNumber)
+                .hasToString("https://gitlab.com/otherns/otherproj/-/merge_requests/1");
+
+        // Project URL with trailing slash
+        assertThat(GitLabRequestLink.resolve(urlOf("https://gitlab.com/nbbrd/heylogs-ext-gitlab/"), "!1"))
+                .returns(urlOf("https://gitlab.com"), GitLabRequestLink::getBase)
+                .returns(java.util.Arrays.asList("nbbrd"), GitLabRequestLink::getNamespace)
+                .returns("heylogs-ext-gitlab", GitLabRequestLink::getProject)
+                .returns(1, GitLabRequestLink::getNumber)
+                .hasToString("https://gitlab.com/nbbrd/heylogs-ext-gitlab/-/merge_requests/1");
+    }
+
 }